requirements.txt
playwright>=1.40.0
openai>=1.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
rich>=13.7.0
pydantic>=2.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
